Add: SendMail for Login and Checkout

+++ ShoesShop/Controllers/ProductController.cs
@@ -15,7 +15,7 @@ namespace ShoesShop.Controllers
         {
             return View();
         }
-        public async Task<IActionResult> Details(int Id)
+        public async Task<IActionResult> Details(long Id)
         {
             if (Id == null)
             {
